+++ sdk/advisor/azure-resourcemanager-advisor/src/main/java/com/azure/resourcemanager/advisor/fluent/models/ResourceRecommendationBaseInner.java
@@ -6,12 +6,14 @@ package com.azure.resourcemanager.advisor.fluent.models;
 
 import com.azure.core.annotation.Fluent;
 import com.azure.core.management.ProxyResource;
+import com.azure.core.management.SystemData;
 import com.azure.json.JsonReader;
 import com.azure.json.JsonToken;
 import com.azure.json.JsonWriter;
 import com.azure.resourcemanager.advisor.models.Category;
 import com.azure.resourcemanager.advisor.models.Impact;
 import com.azure.resourcemanager.advisor.models.ResourceMetadata;
+import com.azure.resourcemanager.advisor.models.Risk;
 import com.azure.resourcemanager.advisor.models.ShortDescription;
 import java.io.IOException;
 import java.time.OffsetDateTime;
@@ -28,6 +30,11 @@ public final class ResourceRecommendationBaseInner extends ProxyResource {
      * The properties of the recommendation.
      */
     private RecommendationProperties innerProperties;
+
+    /*
+     * Azure Resource Manager metadata containing createdBy and modifiedBy information.
+     */
+    private SystemData systemData;
 
     /*
      * The type of the resource.
@@ -57,6 +64,15 @@ public final class ResourceRecommendationBaseInner extends ProxyResource {
      */
     private RecommendationProperties innerProperties() {
         return this.innerProperties;
+    }
+
+    /**
+     * Get the systemData property: Azure Resource Manager metadata containing createdBy and modifiedBy information.
+     * 
+     * @return the systemData value.
+     */
+    public SystemData systemData() {
+        return this.systemData;
     }
 
     /**
@@ -247,6 +263,29 @@ public final class ResourceRecommendationBaseInner extends ProxyResource {
             this.innerProperties = new RecommendationProperties();
         }
         this.innerProperties().withRecommendationTypeId(recommendationTypeId);
+        return this;
+    }
+
+    /**
+     * Get the risk property: The potential risk of not implementing the recommendation.
+     * 
+     * @return the risk value.
+     */
+    public Risk risk() {
+        return this.innerProperties() == null ? null : this.innerProperties().risk();
+    }
+
+    /**
+     * Set the risk property: The potential risk of not implementing the recommendation.
+     * 
+     * @param risk the risk value to set.
+     * @return the ResourceRecommendationBaseInner object itself.
+     */
+    public ResourceRecommendationBaseInner withRisk(Risk risk) {
+        if (this.innerProperties() == null) {
+            this.innerProperties = new RecommendationProperties();
+        }
+        this.innerProperties().withRisk(risk);
         return this;
     }
 
@@ -553,6 +592,8 @@ public final class ResourceRecommendationBaseInner extends ProxyResource {
                 } else if ("properties".equals(fieldName)) {
                     deserializedResourceRecommendationBaseInner.innerProperties
                         = RecommendationProperties.fromJson(reader);
+                } else if ("systemData".equals(fieldName)) {
+                    deserializedResourceRecommendationBaseInner.systemData = SystemData.fromJson(reader);
                 } else {
                     reader.skipChildren();
                 }