+++ sdk/advisor/azure-resourcemanager-advisor/src/main/java/com/azure/resourcemanager/advisor/implementation/RecommendationMetadatasImpl.java
@@ -5,8 +5,6 @@
 package com.azure.resourcemanager.advisor.implementation;
 
 import com.azure.core.http.rest.PagedIterable;
-import com.azure.core.http.rest.Response;
-import com.azure.core.http.rest.SimpleResponse;
 import com.azure.core.util.Context;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.resourcemanager.advisor.fluent.RecommendationMetadatasClient;
@@ -25,25 +23,6 @@ public final class RecommendationMetadatasImpl implements RecommendationMetadata
         com.azure.resourcemanager.advisor.AdvisorManager serviceManager) {
         this.innerClient = innerClient;
         this.serviceManager = serviceManager;
-    }
-
-    public Response<MetadataEntity> getWithResponse(String name, Context context) {
-        Response<MetadataEntityInner> inner = this.serviceClient().getWithResponse(name, context);
-        if (inner != null) {
-            return new SimpleResponse<>(inner.getRequest(), inner.getStatusCode(), inner.getHeaders(),
-                new MetadataEntityImpl(inner.getValue(), this.manager()));
-        } else {
-            return null;
-        }
-    }
-
-    public MetadataEntity get(String name) {
-        MetadataEntityInner inner = this.serviceClient().get(name);
-        if (inner != null) {
-            return new MetadataEntityImpl(inner, this.manager());
-        } else {
-            return null;
-        }
     }
 
     public PagedIterable<MetadataEntity> list() {